update CloudFormation template with lambda:InvokeFunction policy
